git-p4: fix "git p4 sync" after ignored changelist
After a sync/clone, if the very next p4 change is outside the client
view (and therefore "empty" as far as git-p4 is concerned), a
subsequent sync will fail with a message like:

     "fast-import failed: warning: Not updating refs/remotes/p4/master
      (new tip xxxxxx does not contain yyyyyy)"

The bug is caused by discarding the parent commit information
unconditionally after processing a p4 change, whether the change was
committed or not.  This leaves the next non-empty commit disconnected
from its parent, causing fast-import to fail.  This only occurs when
git-p4.keepEmptyCommits is false, but that is the default.

Rename P4Sync.commit() to maybeCommit() and return True if the change
is committed, or False if ignored. Clear P4Sync.initialParent only if
maybeCommit() returns True.

Diagnosed by IanH at https://stackoverflow.com/a/39876288/2033415

Signed-off-by: Evan McLain <git.commit@none-of-yer.biz>