[GSOC] cat-file: fix --batch report changed-type bug
When `--batch` used with `--batch-all-objects`,
with some format atoms like %(objectname), %(rest)
or even no atoms may cause Git exit and report
"object xxx changed type!?".

E.g. `git cat-file --batch="batman" --batch-all-objects`

This is because we did not get the object type through
oid_object_info_extended(), it's composed of two
situations:

1. Since object_info is empty, skip_object_info is
set to true, We skipped collecting the object type.

2. The formatting atom like %(objectname) does not require
oid_object_info_extended() to collect object types.

The correct way to deal with it is to swap the order
of setting skip_object_info and setting typep. This
will ensure that we must get the type of the object
when using --batch.

Helped-by: Jeff King <peff@peff.net>
Signed-off-by: ZheNing Hu <adlternative@gmail.com>